Start converting to uniform REST principles

+++ migrations/2019-10-10-200000_create_players/up.sql
@@ -4,5 +4,5 @@ CREATE TABLE players (
        name_family   VARCHAR(32)  NOT NULL,
        name_given    VARCHAR(32)  NOT NULL,
        date_of_birth DATE         NOT NULL,
-       height_cm     INT          NOT NULL,
-       weight_kg     NUMERIC(3,1) NOT NULL);
+       height_cm     REAL         NOT NULL,
+       weight_kg     REAL         NOT NULL);
